+++ src/components/Details.js
@@ -1,12 +1,14 @@
 import React from 'react';
-import { useParams, Link } from 'react-router-dom';
+import { useParams, useNavigate } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 import './details.css';
-import SearchImg from '../assets/images/searchIcon.png';
+// import SearchImg from '../assets/images/searchIcon.png';
 
 const Details = () => {
   const { id } = useParams();
   const cryptosArray = useSelector((state) => state.stCryptos);
+
+  const navigate = useNavigate();
 
   let data = cryptosArray.filter((obj) => obj.id === id);
   data = data ? data[0] : null; // or undefined
@@ -15,18 +17,13 @@ const Details = () => {
 
   console.log(cryptosArray, id, data);
 
-  // const query = document.getElementById('submit');
-  // query.style.display = 'none';
-
   /* eslint-disable consistent-return */
   const renderData = () => {
     if (data) {
       return (
         <div className="detail_div11">
           <div className="header_lo1111go">
-            <Link to="/">
-              <img src={SearchImg} alt="return" />
-            </Link>
+            <button type="button" onClick={() => navigate(-1)}>Go back</button>
           </div>
           <div className="det_head">
             <div className="det_headName">
